Add example template for tensor loading and argument creation
- Introduced a new Python template file `example.py` in the reproducer templates directory.
- Implemented functions for loading tensors with integrity verification and creating kernel launch arguments from a JSON file.
- Added detailed docstrings for each function to enhance code clarity and usability.
- Included error handling for file operations and argument processing to improve robustness.

This template serves as a foundational example for users to understand how to utilize the reproducer functionality effectively.
